Disable I/O channel buffering
Currently, the I/O channel for client-server communication is buffered,
which means that nothing gets sent to IMMS (including the initial Setup
command) for a while, likely until the internal pipe buffer fills up.
Practically speaking, this means that IMMS often doesn't see any input
until songs have been playing for a while, and it processes large
numbers of songs at once instead of one song at a time. This has two
issues:

(1) For short-lived sessions with just a few songs played, song data is
lost.

(2) A delay before processing a song means that IMMS' database is
temporarily out-of-date. Although this is not a huge problem, it's easy
to correct, so we might as well do that.

Given that reading and acting on data is a fairly cheap operation, turn
off buffering to fix this issue.

This fixes #24.