fix type-punning warning during compile
In the past, we got the following warning:

../immscore/md5.c: In function ‘md5_finish_ctx’:
../immscore/md5.c:109:3: warning: dereferencing type-punned pointer will
break strict-aliasing rules [-Wstrict-aliasing]
   *(md5_uint32 *) &ctx->buffer[bytes + pad] = SWAP (ctx->total[0] <<
   3);

This is type-punning, caused by taking the address of ctx->buffer and
then dereferencing it as an md5_uint32, which can have endianness and
other issues. In addition, it breaks gcc strict aliasing, which can
cause a large host of nasty bugs and vary with optimization levels. Fix
this by using a union of char and md5_uint32 so that the cast is no
longer necessary. As a side-benefit, the code is easier to read, too :).